回复功能及其异常处理

+++ src/main/resources/db/migration/V3__Create_question_table.sql
@@ -1,11 +1,11 @@
 create table question
 (
-	id int auto_increment primary key,
+	id bigint auto_increment primary key,
 	title varchar(50) null,
 	description text null,
 	gmt_create bigint null,
 	gmt_modified bigint null,
-	creator int null,
+	creator bigint null,
 	comment_count int default 0 null,
 	view_count int default 0 null,
 	like_count int default 0 null,
